requirements.txt
streamlit>=1.36,<2
pandas>=2.2
numpy>=1.26
plotly>=5.22
yfinance>=0.2.43
python-dotenv>=1.0
alpaca-py>=0.21
